Fix tone(int,int,int) infinite loop
Explicitly cast the frequency, when passed in as an int, to an
unsigned int.  Verified with snippet:
  tone(D1, (int)1000, 500);
  tone(D1, (unsigned int)1000, 500);
  tone(D1, 1000.0, 500);
  tone(D1, (int)1000);
  tone(D1, (unsigned int)1000);
  tone(D1, 1000.0);

+++ cores/esp8266/Tone.cpp
@@ -72,7 +72,8 @@ void tone(uint8_t _pin, double frequency, unsigned long duration) {
 
 // Fix ambiguous tone() binding when adding in a duration
 void tone(uint8_t _pin, int frequency, unsigned long duration) {
-  tone(_pin, frequency, duration);
+  // Call the unsigned int version of the function explicitly
+  tone(_pin, (unsigned int)frequency, duration);
 }
